fix another implicit float cast

+++ domain-server/src/DomainServer.cpp
@@ -884,10 +884,10 @@ QJsonObject DomainServer::jsonObjectForNode(const SharedNodePointer& node) {
         
         if (!nodeData->getWalletUUID().isNull()) {
             TransactionHash::iterator i = _pendingAssignmentCredits.find(nodeData->getWalletUUID());
-            double pendingCreditAmount = 0;
+            float pendingCreditAmount = 0;
             
             while (i != _pendingAssignmentCredits.end() && i.key() == nodeData->getWalletUUID()) {
-                pendingCreditAmount += i.value()->getAmount() / SATOSHIS_PER_CREDIT;
+                pendingCreditAmount += i.value()->getAmount() / (float) SATOSHIS_PER_CREDIT;
                 ++i;
             }
